Added: Labyrinth start - cell draw ESC now exits the process and starts a dtor of GameDemo
Anonymous inheritance is not a thing in C so it is direct

Defined this pointer and base(value type, not pointer), but that will do.
Got rid of ugly code in c-tors and virtual funcs.

Fixed: artifacts from uninitialized buffer.
destructor free pointers are now pointers, not values.

TODO:\\
The First Game!
Now 'tis possible to make maze gen