Probably the last commit:
Made the game of maze in 3d, true color schemes are way too slow to implement, so I'm switching to D3D for a new project of VeritasEngine based on D3D or GL, because writing drivers to videocard is not what I can do atm.

Although several experiments shall be run for clarification, but for now, this is the end.